Use a whole-word value in a test, to fix an unaligned crash on sparc64.
All our uses in real code are pointers, so it's just this test.
Someone reusing the code as a container for 32-bit values would need to
fix it, but at the moment I don't care.  And non-sparc can do unaligned.

(cherry picked from commit c7cfbf4d2ae3d9c201caf90418e14a6676801716)

+++ tests/kbitr.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "../kbtree.h"
 
-typedef int num;
+typedef intptr_t num;
 
 static num numcmp(num a, num b)
 {
@@ -31,7 +31,7 @@ int main()
                 num l = (k + 1) | 1;
                 if (j != l)
                 {
-                    printf("got %d wanted %d\n", j, l);
+                    printf("got %zd wanted %zd\n", j, l);
                     return 1;
                 }
             }
